# openapi.yaml
openapi: 3.1.0
info:
  title: Puntazo API · programa gratuito
  version: 0.3.0-contract
  description: |
    Contrato de lanzamiento gratuito para programas Sellos y Puntos.
    Cada operación marcada IMPLEMENTED cuenta con código, persistencia, autorización
    y pruebas de contrato/integración dentro del repositorio aislado de preview.
  x-puntazo-slice: FREE-LAUNCH-SELLOS-PUNTOS
  x-puntazo-authority: ACORDADO_PARA_DEMO
  x-puntazo-approved-decisions: [PC-04, PC-09, PC-10, PC-11, PC-13, PC-14]
  x-puntazo-partial-decisions: [PC-12]
  x-puntazo-deferred: [PC-01, PC-03, PC-05, PC-06, PC-07, PC-08]
servers:
  - url: https://puntazo-preview.blakyta3d.duckdns.org/api/v1
    description: Staging VPS; el gateway web es el único servicio público
  - url: http://localhost:8080/v1
    description: Desarrollo local de API
security:
  - bearerAuth: []
...
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      operationId: registerCustomer
      summary: Registra exclusivamente una cuenta cliente final gratuita
      security: []
      parameters:
        - $ref: '#/components/parameters/ClientPlatform'
      x-puntazo-implementation: IMPLEMENTED
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterCustomerRequest' }
      responses:
        '201':
          description: Cliente creado sin suscripción; QR emitido y almacenado sólo como hash
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RegisterCustomerEnvelope' }
        '403': { $ref: '#/components/responses/DemoSignupClosed' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
        '429': { $ref: '#/components/responses/RateLimited' }
...
  /auth/google:
    post:
      tags: [Auth]
      operationId: loginGoogle
      summary: Inicia sesión o crea una cuenta elegida con un ID token de Google verificado
      description: |
        Para una cuenta existente basta `id_token`; `account_type` y `merchant_registration`
        se ignoran y nunca convierten la cuenta. Una identidad nueva debe indicar
        `CLIENTE_FINAL` o `PERSONAL_MARCA`. El alta `PERSONAL_MARCA` provisiona en una
        transacción la marca, membresía PROPIETARIO, sucursal, programa y acceso demo $0.
        Si una identidad Google verificada es nueva y omite `account_type`, no se crea
        usuario, sesión ni recurso comercial: se devuelve `422 ACCOUNT_TYPE_REQUIRED`
        con `details.next_action=SELECT_ACCOUNT_TYPE`. El cliente debe conservar el
        `id_token` sólo en memoria y reenviarlo después de la elección; si Google ya no
        lo valida, debe comenzar una autenticación nueva. El alias temporal
        `POST /auth/google` conserva este mismo error y reglas de selección.
      security: []
      x-puntazo-implementation: IMPLEMENTED
      parameters:
        - $ref: '#/components/parameters/ClientPlatform'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/GoogleAuthRequest' }
      responses:
        '200':
          description: Sesión iniciada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthEnvelope' }
        '403': { $ref: '#/components/responses/DemoSignupClosed' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '422': { $ref: '#/components/responses/GoogleAuthValidationError' }
        '429': { $ref: '#/components/responses/RateLimited' }
...
  /me:
    get:
      tags: [Cuenta]
      summary: Consulta la cuenta y sus membresías activas
      operationId: getCurrentUser
      x-puntazo-implementation: IMPLEMENTED
      responses:
        '200':
          description: Identidad y membresías restaurables en otro dispositivo
          headers:
            ETag:
              description: Versión fuerte para mutaciones con If-Match
              schema: { type: string, pattern: '^"[1-9][0-9]*"$' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CurrentUserEnvelope' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
    patch:
      tags: [Cuenta]
      summary: Actualiza parcialmente el perfil con control de versión
      operationId: updateCurrentUser
      x-puntazo-implementation: IMPLEMENTED
      parameters:
        - name: If-Match
          in: header
          required: true
          schema: { type: string, pattern: '^"[1-9][0-9]*"$' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateAccountRequest' }
      responses:
        '200':
          description: Perfil actualizado
          headers:
            ETag:
              schema: { type: string, pattern: '^"[1-9][0-9]*"$' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CurrentUserEnvelope' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '412':
          description: If-Match ausente, inválido o desactualizado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    delete:
      tags: [Cuenta]
      operationId: anonymizeCurrentUser
      summary: Anonimiza la cuenta y preserva el ledger histórico
      x-puntazo-implementation: IMPLEMENTED
      parameters:
        - name: If-Match
          in: header
          required: true
          schema: { type: string, pattern: '^"[1-9][0-9]*"$' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AnonymizeAccountRequest' }
      responses:
        '202':
          description: Anonimización completada y accesos revocados
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AnonymizationEnvelope' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '409':
          description: La cuenta sigue siendo el último propietario activo de una marca
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
        '412':
          description: If-Match ausente, inválido o desactualizado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  parameters:
    ClientPlatform:
      name: X-Client-Platform
      in: header
      required: true
      description: web usa cookie HttpOnly; native transporta refresh_token en JSON y debe guardarlo en SecureStore
      schema: { type: string, enum: [web, native] }
...
  responses:
    Unauthenticated:
      description: Sesión ausente o inválida
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
...
    Conflict:
      description: Conflicto de email, idempotencia, preview o saldo
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
    ValidationError:
      description: Request válido como JSON pero inválido para el contrato
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
    GoogleAuthValidationError:
      description: |
        Request inválido o identidad Google verificada que requiere elegir el tipo de
        cuenta antes de realizar cualquier escritura o emitir una sesión.
      content:
        application/json:
          schema:
            oneOf:
              - $ref: '#/components/schemas/AccountTypeRequiredErrorEnvelope'
              - $ref: '#/components/schemas/InvalidRequestErrorEnvelope'
    RateLimited:
      description: Límite por IP, email u operador excedido
      headers:
        Retry-After: { schema: { type: integer, minimum: 1 } }
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
...
    DemoSignupClosed:
      description: Altas cerradas o código de demo inválido
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
  schemas:
...
    Timestamp:
      type: string
      format: date-time
    RequestId:
      type: string
      format: uuid
...
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error, request_id]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message]
          properties:
            code:
              type: string
              enum: [INVALID_REQUEST, ACCOUNT_TYPE_REQUIRED, IDENTITY_TOKEN_INVALID, INVITATION_INVALID, INVITATION_EMAIL_MISMATCH, SELF_ROLE_CHANGE_FORBIDDEN, EMAIL_VERIFICATION_REQUIRED, UNAUTHENTICATED, RECENT_AUTH_REQUIRED, FORBIDDEN, NOT_FOUND, EMAIL_EXISTS, RESOURCE_CONFLICT, PROGRAM_TYPE_HAS_BENEFITS, PROGRAM_TYPE_IMMUTABLE, IDEMPOTENCY_CONFLICT, IDEMPOTENCY_IN_PROGRESS, PREVIEW_EXPIRED, PREVIEW_CONSUMED, PREVIEW_CHANGED, PRECONDITION_FAILED, OWNERSHIP_TRANSFER_REQUIRED, INSUFFICIENT_BALANCE, DEMO_SIGNUP_DISABLED, DEMO_ACCESS_DENIED, MEDIA_TOO_LARGE, MEDIA_TYPE_UNSUPPORTED, MEDIA_STORAGE_UNAVAILABLE, RATE_LIMITED, DEPENDENCY_UNAVAILABLE, INTERNAL_ERROR]
            message: { type: string }
            details: { type: object, additionalProperties: true }
        request_id: { $ref: '#/components/schemas/RequestId' }
    AccountTypeRequiredErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error, request_id]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, details]
          properties:
            code: { type: string, const: ACCOUNT_TYPE_REQUIRED }
            message: { type: string }
            details:
              type: object
              additionalProperties: false
              required: [next_action]
              properties:
                next_action: { type: string, const: SELECT_ACCOUNT_TYPE }
        request_id: { $ref: '#/components/schemas/RequestId' }
    InvalidRequestErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error, request_id]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message]
          properties:
            code: { type: string, const: INVALID_REQUEST }
            message: { type: string }
        request_id: { $ref: '#/components/schemas/RequestId' }
...
    AccountType:
      type: string
      enum: [CLIENTE_FINAL, PERSONAL_MARCA]
    User:
      type: object
      additionalProperties: false
      required: [id, email, name, account_type, active, email_verified, auth_version, version, created_at]
      properties:
        id: { type: integer, format: int64 }
        email: { type: string, format: email }
        name: { type: string }
        apellido: { type: string, maxLength: 120 }
        alias: { type: string, maxLength: 80 }
        foto_url: { type: string, format: uri, maxLength: 2048 }
        account_type: { $ref: '#/components/schemas/AccountType' }
        active: { type: boolean }
        email_verified: { type: boolean }
        auth_version: { type: integer, minimum: 1 }
        version: { type: integer, minimum: 1 }
        created_at: { $ref: '#/components/schemas/Timestamp' }
    UpdateAccountRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        nombre: { type: string, minLength: 1, maxLength: 120 }
        apellido: { type: [string, 'null'], maxLength: 120 }
        alias: { type: [string, 'null'], maxLength: 80 }
...
    AnonymizeAccountRequest:
      type: object
      additionalProperties: false
      required: [confirmacion]
      properties:
        confirmacion: { type: string, const: BAJA }
    Anonymization:
      type: object
      additionalProperties: false
      required: [id_solicitud, estado, acceso_revocado, ledger_preservado, requested_at]
      properties:
        id_solicitud: { type: string, format: uuid }
        estado: { type: string, const: COMPLETADA }
        acceso_revocado: { type: boolean, const: true }
        ledger_preservado: { type: boolean, const: true }
        requested_at: { $ref: '#/components/schemas/Timestamp' }
    AnonymizationEnvelope:
      type: object
      additionalProperties: false
      required: [data, request_id]
      properties:
        data: { $ref: '#/components/schemas/Anonymization' }
        request_id: { $ref: '#/components/schemas/RequestId' }
    Membership:
      type: object
      additionalProperties: false
      required: [brand_id, brand_name, role, branch_ids, active]
      properties:
        brand_id: { type: integer, format: int64 }
        brand_name: { type: string }
        role: { type: string, enum: [PROPIETARIO, ADMINISTRADOR, OPERADOR] }
        branch_ids:
          type: array
          items: { type: integer, format: int64 }
        active: { type: boolean }
    Session:
      type: object
      additionalProperties: false
      required: [access_token, token_type, expires_in]
      properties:
        access_token: { type: string }
        refresh_token:
          type: string
          description: Presente sólo para native; web recibe la cookie puntazo_refresh HttpOnly, Secure y SameSite=Strict
        token_type: { type: string, const: Bearer }
        expires_in:
          type: integer
          const: 900
          description: Vida del access token en segundos; el refresh token rota en cada uso
    AuthData:
      type: object
      additionalProperties: false
      required: [session, user]
      properties:
        session: { $ref: '#/components/schemas/Session' }
        user: { $ref: '#/components/schemas/User' }
    AuthEnvelope:
      type: object
      additionalProperties: false
      required: [data, request_id]
      properties:
        data: { $ref: '#/components/schemas/AuthData' }
        request_id: { $ref: '#/components/schemas/RequestId' }
    RegisterCustomerData:
      type: object
      additionalProperties: false
      required: [user, verification_required]
      properties:
        user: { $ref: '#/components/schemas/User' }
        session: { $ref: '#/components/schemas/Session' }
        verification_required: { type: boolean }
    RegisterCustomerEnvelope:
      type: object
      additionalProperties: false
      required: [data, request_id]
      properties:
        data: { $ref: '#/components/schemas/RegisterCustomerData' }
        request_id: { $ref: '#/components/schemas/RequestId' }
...
    RegisterCustomerRequest:
      type: object
      additionalProperties: false
      required: [email, password, name]
      properties:
        email: { type: string, format: email, maxLength: 254 }
        password:
          type: string
          format: password
          minLength: 10
          maxLength: 72
          description: Máximo 72 bytes en UTF-8 por compatibilidad con bcrypt
        name: { type: string, minLength: 1, maxLength: 120 }
...
    GoogleAuthRequest:
      type: object
      additionalProperties: false
      required: [id_token]
      properties:
        id_token: { type: string, minLength: 1 }
        account_type: { $ref: '#/components/schemas/AccountType' }
        merchant_registration: { $ref: '#/components/schemas/GoogleMerchantRegistration' }
    GoogleMerchantRegistration:
      type: object
      additionalProperties: false
      required: [brand_name, branch_name, program_type]
      properties:
        brand_name: { type: string, minLength: 1, maxLength: 120 }
        branch_name: { type: string, minLength: 1, maxLength: 120 }
        branch_address: { type: string, maxLength: 300 }
        branch_locality: { type: string, maxLength: 120 }
        branch_province: { type: string, maxLength: 120 }
        branch_postal_code: { type: string, maxLength: 20 }
        branch_latitude: { type: number, format: double, minimum: -90, maximum: 90, description: Debe enviarse junto con branch_longitude }
        branch_longitude: { type: number, format: double, minimum: -180, maximum: 180, description: Debe enviarse junto con branch_latitude }
        program_type: { type: string, enum: [SELLOS, PUNTOS] }
...
    CurrentUser:
      type: object
      additionalProperties: false
      required: [user, memberships, onboarding_complete]
      properties:
        user: { $ref: '#/components/schemas/User' }
        memberships:
          type: array
          items: { $ref: '#/components/schemas/Membership' }
        onboarding_complete: { type: boolean }
    CurrentUserEnvelope:
      type: object
      additionalProperties: false
      required: [data, request_id]
      properties:
        data: { $ref: '#/components/schemas/CurrentUser' }
        request_id: { $ref: '#/components/schemas/RequestId' }
...